+++ src/main/java/uz/ccrew/flightmanagement/repository/UserRepository.java
@@ -4,6 +4,6 @@ import uz.ccrew.flightmanagement.entity.User;
 
 import java.util.Optional;
 
-public interface UserRepository extends BasicRepository<User, Integer> {
+public interface UserRepository extends BasicRepository<User, Long> {
     Optional<User> findByLogin(String login);
 }
